requirements.txt
django=5.0.3
django_rest_framework=3.14.0
django-cors-headers=4.3.1
psycopg2=2.9.9 